+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.2.0
+
+- Adds `()` as a langauge feature
+- Adds `return()` as a language feature
+- Adds `function()` as a language feature
+
 ## 0.1.0
 
 The initial version!